[chore] #5 register.ts login.ts 를 auth.ts 로 통합
인증 관련 api 한 곳에서 관리
