+++ src/common/api/auth.ts
@@ -1,7 +1,10 @@
-import { AxiosError } from "axios";
 import apiClient from ".";
 
 interface LoginResult {
+  result: "success" | "fail";
+}
+
+interface RegisterResult {
   result: "success" | "fail";
 }
 
@@ -38,5 +41,33 @@ export const login = async ({
     // }
 
     // throw error;
+  }
+};
+
+export const register = async ({
+  email,
+  password,
+}: {
+  email: string;
+  password: string;
+}): Promise<RegisterResult> => {
+  try {
+    const response = await apiClient.post("/auth/signup", {
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: {
+        email: email,
+        password: password,
+      },
+    });
+
+    return {
+      result: "success",
+    };
+  } catch (error) {
+    return {
+      result: "fail",
+    };
   }
 };
